% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_mysql_sid.R
\name{mysql_operations}
\alias{mysql_operations}
\title{A MySQL Operation}
\usage{
mysql_operations(query, record)
}
\arguments{
\item{select, }{Select records of musician from the database .}

\item{insert, record}{,Insert records into musican table where insert is string and record is a list of values.}
}
\description{
This function allows you to select records from the table and to insert records to the music table.
}
\examples{
mysql_operations('select') for selecting records
mysql_operations('insert',record) for inserting records into music table
}
